requirements.txt
openai
httpx
python-pptx
streamlit
python-dotenv
requests
